pyproject.toml
[tool.poetry]
name = "fastsso"
version = "1.0.1a0"
description = ""
authors = ["Alexandre Meline"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.104.1"}
python-keycloak = "^3.3.0"
pydantic = "^2.4.2"
uvicorn = "^0.24.0.post1"


[tool.poetry.group.test.dependencies]
pytest = "^7.4.3"
requests-mock = "^1.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
